Fix ContextServer
And random fixes
