+++ Project/src/fr/umlv/java/models/Commands.java
@@ -1,15 +1,17 @@
 package fr.umlv.java.models;
 
+import fr.umlv.java.models.context.ContextClient;
+
 import java.util.ArrayDeque;
 
 public class Commands {
     private final int CAPACITY = 10;
     private final static Object lock = new Object();
     private final ArrayDeque<Message> queue = new ArrayDeque<>(CAPACITY);
-    private final Context context;
+    private final ContextClient contextClient;
 
-    public Commands(Context context) {
-        this.context = context;
+    public Commands(ContextClient contextClient) {
+        this.contextClient = contextClient;
     }
     public void transferMessage(Message msg) throws InterruptedException {
         synchronized (lock) {
@@ -23,7 +25,7 @@ public class Commands {
     public void retrieveCommand() {
         synchronized (lock) {
             while(!queue.isEmpty()) {
-                context.queueMessage(queue.pop());
+                contextClient.queueMessage(queue.pop());
             }
             lock.notify();
         }
