+++ Project/src/fr/umlv/java/models/context/ContextClient.java
@@ -1,5 +1,7 @@
-package fr.umlv.java.models;
+package fr.umlv.java.models.context;
 
+import fr.umlv.java.models.ConnectionStatut;
+import fr.umlv.java.models.Message;
 import fr.umlv.java.readers.ListStringReader;
 import fr.umlv.java.readers.Reader;
 
@@ -12,7 +14,9 @@ import java.nio.charset.StandardCharsets;
 import java.util.ArrayDeque;
 import java.util.List;
 
-public class Context {
+public class ContextClient {
+
+    static private final Charset UTF_8 = StandardCharsets.UTF_8;
     static private int BUFFER_SIZE = 10_000;
     private final SelectionKey key;
     private final SocketChannel sc;
@@ -27,9 +31,7 @@ public class Context {
     private byte currentOpCode = -1;
     private String serverName;
 
-    static private final Charset UTF_8 = StandardCharsets.UTF_8;
-
-    public Context(SelectionKey key, Boolean logged, String login) {
+    public ContextClient(SelectionKey key, Boolean logged, String login) {
         this.key = key;
         this.sc = (SocketChannel) key.channel();
         this.logged = logged;
@@ -100,7 +102,7 @@ public class Context {
             return;
         }
         var login_buffer = UTF_8.encode(msg.getLogin());
-        var msg_buffer = UTF_8.encode(msg.getTexte());
+        var msg_buffer = UTF_8.encode(msg.getText());
         var servername_buffer = UTF_8.encode(serverName);
         if (bufferOut.remaining() < login_buffer.remaining() + msg_buffer.remaining() + Integer.BYTES*2) {
             return;
